Added Year field to registration form
The logic for SignInController was modified, both to accomodate the
Year field and to fix a few bugs when switching between the sign-in and
registration modes. Also, creating a new user also creates a new
Student in the database now. To assist with this, some Student columns
are now blank by default to handle new accounts. Also, a bug with
FollowingModel was fixed.