Added config file reader to load database password
The config file	is src/settings.conf. It can be	used to specify the
database password. The file should contain a line starting with
"db_password=" (without quotes), followed by the password (without a
space after the equals sign). If the config file cannot be found or
does not contain the database password, the app will instead use the
default password, which is "12345" (without quotes).