Added helper script and support for Faculty registration
Users can choose to register as either students or faculty members.
This determines the value that is set for the new User column user_type
(which stores a 0 for students and a 1 for faculty members). Also, the
new Faculty table is currently very similar to the Student table.

I also added a helper script that will assist in future changes made
to the .sql files. I also removed unneeded exports.