Added chat and unfriending abilities to Community page
A user must be friends with someone in order to send messages to them.
If a user isn't friends with someone, then they won't see any messages
sent to them from that person, nor will they be able to send anything
to that person. This should prevent the user from seeing unwanted
messages and spam.

A new database table was added to allow for this functionality. Also,
the sql_scripts/adjust_restaurants_and_events.py script was modified
to remove occurrences of "AUTO_INCREMENT=".