Added Event Calendar page and modified Home page
The Event Calendar displays the events of the current month and allows
the current user to confirm or deny their attendance for the month's
events (assuming those events will occur today or after today). Users
can go also go forward to future months or backward to previous months.

To assist in switching to the Event Calendar, most of the pages for
signed-in users now subclass AbstractControllerWithNav (which contains
most of the code originally in AbstractController). The Event Calendar
page, however, subclasses AbstractController (which no longer has the
code for the navigation bar) since the Event Calendar page doesn't have
a navigation bar.

I also modified the Home page to allow users to deny their attendance
(in addition to confirming their attendance) to various upcoming
events. Also, The home page only lists the upcoming events of the
current month.