Made follower code more efficient and fixed event text
I modified the code that determines if a user is a follower of the
current user (when listing the events that followers have attended or
will attend). It is more efficient now.

I also fixed the text listing the events that followers have attended
or will attend. If a follower hasn't confirmed their attendance for any
events occurring this month, then a sentence will be printed stating
this.

I also fixed several mistakes in various comments.